CASMCMS-8927: Add option to IMS export tool to estimate size of exported data
